start to move aggressively away from native boxen modules
for all nonessential software.

boxen modules have proven to be slowly updated, with pull requests
frequently languishing.  also keeping them updated requires a huge
amount of personal sysadmin work since it results in normal
consumer-space applications being versioned.

with this, we are beginning to move towards using homebrew packages as
much as possible for system stuff, and homebrew-cask for application
stuff.  the plan is this can still be managed via boxen puppet files,
but will use the more reliable and frequently updated package
management sources as much as possible.

there are still a number of todos here, but this is a good start
towards giving people a better way to handle installing applications.
feels clean, man.

notable that this is going to result in a lot of annoying output until
phinze/puppet-brewcask#6 is resolved, so it’s not suitable to migrate
anyone else in the org until then.